Step 9 - Ok

+++ front/js/cart.js
@@ -1,55 +1,60 @@
 
 
 let cart = []
-let quantityProductTotal = 0
-let totalProductPrice = 0
+productsCart()
 
-fetch("http://localhost:3000/api/products/")
-   .then(data => data.json())
-   .then(jsonListProduct => {
-      for (product in localStorage)
-         if (localStorage.hasOwnProperty(product)) {
-            // console.log(product) //#########
-            product = JSON.parse(localStorage[product])[0]
-            // console.log(product) //#########
-            // console.log(product.reference) //#########
-            let referenceId = product.reference.split(" ")[0]
-            // console.log(referenceId) //#########
-            let referenceColor = product.color.split(" ")
-            // console.log(referenceColor) //#########
-            let cartProductQuantity = product.quantity
-            // console.log("Voici la quantité : " + cartProductQuantity) //#########
+function productsCart(bool = true) {
+   let quantityProductTotal = 0
+   let totalProductPrice = 0
+   fetch("http://localhost:3000/api/products/")
+      .then(data => data.json())
+      .then(jsonListProduct => {
+         for (product in localStorage)
+            if (localStorage.hasOwnProperty(product)) {
+               // console.log(product) //#########
+               product = JSON.parse(localStorage[product])[0]
+               // console.log(product) //#########
+               // console.log(product.reference) //#########
+               let referenceId = product.reference.split("_")[0]
+               // console.log(referenceId) //#########
+               let referenceColor = product.color.split("_")
+               // console.log(referenceColor) //#########
+               let cartProductQuantity = product.quantity
+               // console.log("Voici la quantité : " + cartProductQuantity) //#########
 
-            for (product_ in jsonListProduct) {
-               let cartProductPrice = jsonListProduct[product_].price
+               for (product_ in jsonListProduct) {
+                  let cartProductPrice = jsonListProduct[product_].price
 
-               if (referenceId == jsonListProduct[product_]._id) {
-                  quantityProductTotal += parseInt(cartProductQuantity)
-                  totalProductPrice += parseInt(cartProductPrice) * parseInt(cartProductQuantity)                  
-                  cart.push(jsonListProduct[product_])
-                  inserProductHtml(jsonListProduct[product_], referenceId, referenceColor, cartProductQuantity)
+                  if (referenceId == jsonListProduct[product_]._id) {
+                     quantityProductTotal += parseInt(cartProductQuantity)
+                     totalProductPrice += parseInt(cartProductPrice) * parseInt(cartProductQuantity)
+                     cart.push(jsonListProduct[product_])
+                     bool && inserProductHtml(jsonListProduct[product_], referenceId, referenceColor, cartProductQuantity)
+                  }
                }
             }
-         }
 
-         let modifierQuantity = document.getElementsByClassName("itemQuantity")
-         for (modifierToQuantity in modifierQuantity)
-         if (modifierQuantity.hasOwnProperty(modifierToQuantity)){
-            console.log(modifierQuantity[modifierToQuantity])
-            modifierQuantity[modifierToQuantity].addEventListener("change", addModifierQuantity)
-         }
+         if (bool) {
+            let modifierQuantity = document.getElementsByClassName("itemQuantity")
+            for (modifierToQuantity in modifierQuantity)
+               if (modifierQuantity.hasOwnProperty(modifierToQuantity)) {
+                  console.log(modifierQuantity[modifierToQuantity])
+                  modifierQuantity[modifierToQuantity].addEventListener("change", addModifierQuantity)
+               }
 
-         let deletedProductCart = document.getElementsByClassName("deleteItem")
-         for (deletedProductToCart in deletedProductCart)
-         if (deletedProductCart.hasOwnProperty(deletedProductToCart)){
-            // console.log(deletedProductCart[deletedProductToCart])
-            deletedProductCart[deletedProductToCart].addEventListener("click", deletedProduct)
+            let deletedProductCart = document.getElementsByClassName("deleteItem")
+            for (deletedProductToCart in deletedProductCart)
+               if (deletedProductCart.hasOwnProperty(deletedProductToCart)) {
+                  // console.log(deletedProductCart[deletedProductToCart])
+                  deletedProductCart[deletedProductToCart].addEventListener("click", deletedProduct)
+               }
          }
 
          inserTotalsHtml(quantityProductTotal, totalProductPrice)
          console.log("Voici la quantité TT du panier : " + quantityProductTotal) //#########
          console.log("Voici le prix TT du panier : " + totalProductPrice) //#########
-      });
+      })
+}
 
 function inserProductHtml(cartProduct, referenceId, referenceColor, cartProductQuantity){
    cart__items.innerHTML += 
@@ -77,8 +82,8 @@ function inserProductHtml(cartProduct, referenceId, referenceColor, cartProductQ
 }
 
 function inserTotalsHtml(quantityTotal, priceTotal){
-   totalQuantity.innerHTML += quantityTotal
-   totalPrice.innerHTML += priceTotal
+   totalQuantity.innerHTML = quantityTotal
+   totalPrice.innerHTML = priceTotal
 }
 
 function addModifierQuantity(e) {
@@ -86,8 +91,10 @@ function addModifierQuantity(e) {
    const color = e.target.closest("article.cart__item").dataset["color"]
 
    // console.log(id, color)
-   let product = JSON.parse(localStorage[id + " " + color])
+   let product = JSON.parse(localStorage[id + "_" + color])
    product[0].quantity = e.target.value
+   localStorage[id + "_" + color] = JSON.stringify(product)
+   productsCart(false)
    console.log(product)
 }
 
@@ -96,7 +103,8 @@ function deletedProduct(e) {
    const color = e.target.closest("article.cart__item").dataset["color"]
 
    // console.log(id, color)
-   delete localStorage[id + " " + color]
+   delete localStorage[id + "_" + color]
    e.target.closest("article.cart__item").remove()
+   productsCart(false)
 }
